Remove spell hunger
Spell hunger was meant to serve as a backstop limitation to medium term
use of magic and the use of magic in conjunction with certain god
abilities. However, except for very high level spells cast with low
intelligence it was rarely relevant in the tactical scale. The spell
hunger mechanic comes with a design cost: it necessitates a wide
difference in food clock costs between magic and non-magic combat
skills, leading to chunk eating and the attendantly loose food clock.

Removing spell hunger paves the way for removing chunk eating, coming in
the next commit.